modify migration and add new migration for MCHAT

+++ database/migrations/2024_09_20_083302_create_milestone_questions_table.php
@@ -4,20 +4,21 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('milestone_questions', function (Blueprint $table) { // plural table name
-            $table->id();
-            $table->unsignedBigInteger('milestone_checklist_id');
-            $table->foreign('milestone_checklist_id')->references('id')->on('milestone_checklists')->onDelete('cascade');
-            $table->text('question');
-            $table->timestamps();        
-        });
+        if (!Schema::hasTable('milestone_questions')) {
+            Schema::create('milestone_questions', function (Blueprint $table) { // plural table name
+                $table->id();
+                $table->unsignedBigInteger('milestone_checklist_id');
+                $table->foreign('milestone_checklist_id')->references('id')->on('milestone_checklists')->onDelete('cascade');
+                $table->text('question');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
