+++ database/migrations/2024_12_04_162412_create_growth_records_table.php
@@ -10,16 +10,18 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::create('growth_records', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('child_id');
-            $table->double('height', 8, 2)->nullable();
-            $table->double('weight', 8, 2)->nullable();
-            $table->date('recorded_at')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('growth_records')) {
+            Schema::create('growth_records', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('child_id');
+                $table->double('height', 8, 2)->nullable();
+                $table->double('weight', 8, 2)->nullable();
+                $table->date('recorded_at')->nullable();
+                $table->timestamps();
 
-            $table->foreign('child_id')->references('id')->on('children')->onDelete('cascade');
-        });
+                $table->foreign('child_id')->references('id')->on('children')->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
